+++ src/index.ts
@@ -12,7 +12,7 @@ export { anyNumber, integer, number, positiveInteger, positiveNumber } from './n
 export { bigint } from './numbers';
 export { exact, inexact, object, pojo } from './objects';
 export { nonEmptyString, regex, string } from './strings';
-export { email, httpsUrl, url, uuid, uuidv1, uuidv4 } from './strings';
+export { email, httpsUrl, url, uuid, uuidv1, uuidv4, nanoid } from './strings';
 export { decimal, hexadecimal, numeric } from './strings';
 export { either, enum_, oneOf, select, taggedUnion } from './unions';